Disable delete_by_query test for 0.18.7

+++ eson-http/test/http/index_test.rb
@@ -80,21 +80,23 @@ context 'HTTP client quick api' do
     asserts("number of docs") { topic["docs"].length }.equals(2)
   end
 
-  context "delete_by_query" do
-    setup do
-      client.index :index => "delete_by_query",
-                   :type => "foo",
-                   :doc => {:foo => :bar}
-      client.index :index => "delete_by_query",
-                   :type => "foo",
-                   :doc => {:foo => :bar}
-      client.refresh :index => "delete_by_query"
-
-      client.delete_by_query :index => "delete_by_query",
-                             :query => { :match_all => {} }
+  unless ENV["ES_VERSION"] && ENV["ES_VERSION"] < "0.19.0"
+    context "delete_by_query" do
+      setup do
+        client.index :index => "delete_by_query",
+                     :type => "foo",
+                     :doc => {:foo => :bar}
+        client.index :index => "delete_by_query",
+                     :type => "foo",
+                     :doc => {:foo => :bar}
+        client.refresh :index => "delete_by_query"
+    
+        client.delete_by_query :index => "delete_by_query",
+                               :query => { :match_all => {} }
+      end
+    
+      asserts("no doc left") { client.search(:index => "delete_by_query")["hits"]["total"] }.equals(0)
     end
-
-    asserts("no doc left") { client.search(:index => "delete_by_query")["hits"]["total"] }.equals(0)
   end
 end
